Add ESM import option
Switch from "main" to "exports" in package.json. All import paths
documented in the README are added under "exports". A separate .mjs file is defined as the ESM import path for 'obliterator', which merely imports and then re-exports the files. Finally,
"type": "commonjs" is explicitly specified to improve clarity, although
it's not strictly necessary.

 - require('obliterator') works as before.
 - require('obliterator/iterator') and the rest will work as before.
 - ESM imports of 'obliterator' work correctly.
 - ESM imports of 'obliterator/iterator' and other subpackages may not
   work.

This is intended to be a _non-breaking_ change for all users, but given
the widespread adoption of obliterator and the finnicky nature of the
import landscape, it should be tested carefully.

I saw some files, such as 'take-into.js', that were not listed as
importable in the README. I left these out of the exports since they're
not documented, meaning any `require('obliterator/take-into')` will
cease to work. If this is not desired, I'd appreciate a full list of
supported entry points.